% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brushthat.R
\docType{package}
\name{brushthat-package}
\alias{brushthat}
\alias{brushthat-package}
\title{brushthat: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/brushthat}
  \item \url{http://krlmlr.github.io/brushthat}
  \item Report bugs at \url{https://github.com/krlmlr/brushthat/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

}
