% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{shine}
\alias{shine}
\title{Displays a gadget that interacts with testthat}
\usage{
shine(pkg = ".")
}
\arguments{
\item{pkg}{The package to test}
}
\description{
Shows test results, allows navigating to failures and re-run.
}
